Commit lots of pending changes
(Go ahead and yell at me for doing one huge commit, but I'm not sorting
out this mess that has built up over time)

+++ src/logic/PlaylistModel.cpp
@@ -42,7 +42,7 @@ PlaylistModel::~PlaylistModel()
 // Node callbacks
 void PlaylistModel::beginInsertNode(PlaylistNode *parent, size_t index)
 {
-    qDebug("insert on %p: %lu", parent, index);
+    qDebug("insert on %p: %lu", parent, (unsigned long)index);
     beginInsertRows(createIndexForNode(parent), (int)index, (int)index);
 }
